corrigindo artistController.js

+++ src/controllers/artistController.js
@@ -6,20 +6,21 @@ export const artistController = {
             const allArtists = await artistService.getAllArtists();
 
             if(allArtists.length === 0) {
-                res.status(200).json(
+                res.status(404).json(
                     {
-                        'success': true,
+                        'success': false,
                         'data': 'Não há nenhum artista no banco de dados'
                     }
                 );
-            } else {
-                res.status(200).json(
-                    {
-                        'success': true, 	
-                        'data': allArtists,
-                    }
-                );
+                return;
             }
+            
+            res.status(200).json(
+                {
+                    'success': true, 	
+                    'data': allArtists,
+                }
+            );
         } catch(err) {
             console.error(`Erro ao recuperar todos os artistas: ${err.message}`);
             res.status(500).json(
@@ -38,20 +39,21 @@ export const artistController = {
         try {
             const artistById = await artistService.getArtistById(id);
             if(artistById.length === 0){
-                res.status(200).json(
+                res.status(404).json(
                     {
                         'sucess': true,
                         'data': 'Não há nenhum artista com o id especificado no banco de dados'
                     }
                 );
-            } else {
-                res.status(200).json(
-                    {
-                        'success': true,
-                        'data': artistById,
-                    }
-                );
+                return;
             }
+
+            res.status(200).json(
+                {
+                    'success': true,
+                    'data': artistById,
+                }
+            );
         } catch(err) {
             console.error(`Erro ao recuperar artista com id ${id}: ${err.message}`);
             res.status(500).json(
@@ -98,6 +100,16 @@ export const artistController = {
 
         try {
             const outdatedArtist = await artistService.getArtistById(id);
+            if(outdatedArtist.length === 0 || !outdatedArtist[0]) {
+                res.status(404).json(
+                    {
+                        'sucess': false,
+                        'data': 'Não há nenhum artista com o id especificado no banco de dados'
+                    }
+                );
+                return;
+            }
+
             const { name } = req.body;
             const nameToUse = name || outdatedArtist[0].name;
             const { biography } = req.body;
@@ -134,6 +146,18 @@ export const artistController = {
         const { id } = req.params;
 
         try {
+            const artistById = await artistService.getArtistById(id);
+
+            if(artistById.length === 0 || !artistById[0]) {
+                res.status(404).json(
+                    {
+                        'sucess': false,
+                        'data': 'Não há nenhum artista com o id especificado no banco de dados'
+                    }
+                );
+                return;
+            }
+
             const deletedArtist = await artistService.deleteArtist(id);
             res.status(200).json(
                 {
